Allow missing authentication feature
This feature allows requests with missing autentication to succeed if
NODE_ENV=dev and UNSAFE_ALLOW_MISSING_AUTHENTICATION is enabled in .env.
This is to be used during development so that authentication can be
added to the API without breaking sensor and app not yet supporting
authentication.

Other changes:
- Documented all .env options in docs/.env.md.
- Fixed authenticateJWT to save decoded token in res.locals instead of
req.headers that errored.
- Added authentication in swagger docs, with padlock icon (security:) on
all routes except login and user creation. The try it out button can
be used both with and without clicking the padlock and pasting a jwt
token.